商品规则属性spu管理

+++ ly-item/ly-item-service/src/main/java/com/leyou/item/controller/CategoryController.java
@@ -23,7 +23,7 @@ public class CategoryController {
     CategoryService categoryService;
 
     @RequestMapping("/list")
-    public List<Category> queryList(@RequestParam("pid")Integer pid){
+    public List<Category> queryList(@RequestParam("pid")Long pid){
         return categoryService.queryByPid(pid);
     }
